add Detail dan CookTime
