--- app/src/main/java/com/example/cooktifyapp/navigation/favorite/Favorite.kt
@@ -1,13 +1,0 @@
-package com.example.cooktifyapp.navigation.favorite
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.cooktifyapp.R
-
-class Favorite : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_favorite)
-
-    }
-}
