+++ app/src/main/java/com/example/cooktifyapp/view/cooktime/CookTimeActivity.kt
@@ -1,4 +1,4 @@
-package com.example.cooktifyapp.detail
+package com.example.cooktifyapp.view.cooktime
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -6,12 +6,17 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.example.cooktifyapp.R
+import com.example.cooktifyapp.databinding.ActivityCookTimeBinding
 
-class Detail : AppCompatActivity() {
+class CookTimeActivity : AppCompatActivity() {
+
+    private lateinit var binding:ActivityCookTimeBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        binding = ActivityCookTimeBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_detail)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
