+++ app/src/main/java/com/example/cooktifyapp/view/navigation/favorite/Favorite.kt
@@ -1,11 +1,13 @@
-package com.example.cooktifyapp
+package com.example.cooktifyapp.view.navigation.favorite
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.cooktifyapp.R
 
-class MainActivity : AppCompatActivity() {
+class Favorite : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_favorite)
+
     }
 }
