+++ app/src/main/java/com/example/cooktifyapp/view/navigation/favorite/FavoriteFragment.kt
@@ -1,4 +1,4 @@
-package com.example.cooktifyapp.navigation.favorite
+package com.example.cooktifyapp.view.navigation.favorite
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
