+++ app/src/main/res/layout/activity_on_boarding.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".onBoarding.OnBoarding">
+    tools:context=".view.onBoarding.OnBoarding">
 
     <ImageView
         android:id="@+id/ivLogo"
