+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".onBoarding.Register">
+    tools:context=".view.onBoarding.Register">
 
     <ImageView
         android:id="@+id/ivBg"
