+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".settings.SettingsActivity">
+    tools:context=".view.settings.SettingsActivity">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/cvTheme"
