+++ app/src/main/res/layout/fragment_favorite.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".navigation.favorite.Favorite">
+    tools:context=".view.navigation.favorite.Favorite">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/top_bar"
