+++ app/src/main/res/layout/fragment_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".navigation.profile.ProfileFragment">
+    tools:context=".view.navigation.profile.ProfileFragment">
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
